Remove QueryResult in favor of returning csv from execute_query
To simplify things for now (and to take the route that is most likely in
the long run), I'm serializing to rows sooner in the process, instead of
converting to cols (QueryResult) and then back to rows again, which is
how 99% of user consumption will be.

Using csv and serde serializing (with a placeholder struct to make
rows).

In the future, have to add labels to dims, or replace id (or allow
properties and such)