Replace hashmaps with bitshift indexing for group by
Previously, dim indexes were used as the key in a hashmap which held the
row being aggregrated.

This was slow because hashing took a long time, and because of the
allocation of many small vecs to hold the row of aggregations.

Now, instead, aggregations are held in a Vec which is indexed by the
bit-packed dim values.

So basically, an index is created by bit packing the releveant dim
members together, and that number should be less than u64. This allows
the value at that index to be the holder of the aggregation (or count),
this direct indexing into the vec which is very fast.

The other reason it's fast is that aggregations are held in one vec
(even if it's long), so you don't have to allocate a new vec every time
there's a new dimension index, which was the previous implementation.

The tuples are materialized only at the end of the query process.

Performance is now on par with Monetdb for a basic query. Could easily
be faster if dims were dictionary encoded, because i.e. having the years
bit-packed creates larger aggregation vecs.