[all] Backend trait refactor
I named it `Backend` instead of `Database` to be a bit more generic. It
could plausibly be an in-memory key-value store, or even a call out to
an api.

The `Backend` trait has one method:
- `exec_sql`

This is a first pass, to help determine whether this is the right level
of interface (for example, should we feed it a Query instead of sql
(I decided no, so that at this moment it's easier to have a generic sql
generator)).

The trait may change in the future!

(note, the `Backend` trait is in `tesseract_core`)

After defining the trait, I've broken out the clickhouse logic into its
own crate. Although there is not much logic, it defines the boundary of
backend-specific implementation clearly, and it will be easier to see what
needs to be done to implement a new backend.

The rest is just moving stuff around in the server to accomodate the
above changes.

For future backend implementers, note that as of this commit, only one
db is allowed at a time.

I've also left a stub for a mysql db backend.