Use trait object for databases
Was using a generic constraint, but should be using a trait object to
allow one or the other in the same field of the AppState struct. This
also allows having multiple databases now.

Some notes:

casting to trait object:
https://stackoverflow.com/questions/38294911/how-do-i-cast-a-literal-value-to-a-trait-object

it needs to be safe to send between threads, so add trait bounds
Send + Sync.

https://users.rust-lang.org/t/sending-trait-objects-between-threads/2374

it needs to be clonable to move into the closure that is
used to initialize actix-web, so there's a litle boilerplate
to implement https://users.rust-lang.org/t/solved-is-it-possible-to-clone-a-boxed-trait-object/1714/4