[core] fix bug in captions for non-parents
In making captions available for parents, I didn't properly pare down
the captions that might come from parents when parents is false.

A user might do this, or a bug in logic layer (which I think is the
immediate cause).

So I should check if logic layer is passing through extra captions even
when parents is false.